Make Uuid generator thread_local.
This fixes a race condition.
